+++ client/user/controllers/paymentMethodsCtrl.ng.js
@@ -1,24 +1,25 @@
-angular.module("food-coop").controller("UserPaymentMethodCtrl", function($scope, paymentMethods, $mdToast){
+angular.module("food-coop").controller("UserPaymentMethodCtrl", function($scope, $reactive, paymentMethods, $mdToast){
+  $reactive(this).attach($scope);
   if (paymentMethods) {
     
-    $scope.paymentMethods = angular.copy(paymentMethods)
+    this.paymentMethods = angular.copy(paymentMethods)
     
-    _.map($scope.paymentMethods, function(p) {
-      p.delete = function(cb) {
-        $scope.call('deletePaymentMethod', p.token, cb)
+    _.map(this.paymentMethods, (p) => {
+      p.delete = (cb) => {
+        this.call('deletePaymentMethod', p.token, cb)
         
-        $scope.paymentMethods = _.reject($scope.paymentMethods, function(pay) {
+        this.paymentMethods = _.reject(this.paymentMethods, function(pay) {
           return pay.token === p.token;
         });
       }
-      p.makeDefault = function(cb) {
-        $scope.call('makeDefaultPaymentMethod', p.token, cb)
+      p.makeDefault = (cb) => {
+        this.call('makeDefaultPaymentMethod', p.token, cb)
       }
       return p
     });
   }
 
-  $scope.callback = function(result) {
+  this.callback = function(result) {
     if (result.token) {
       $mdToast.show(
         $mdToast.simple().content("Success!").position('bottom left').hideDelay(4000)