+++ imports/ui/components/eventDetails/eventDetails.js
@@ -1,8 +1,9 @@
 // import angular from 'angular'
 import uiRouter from 'angular-ui-router'
 import ngMaterial from 'angular-material'
-import {name as eventOrderForm} from '../eventOrderForm/eventOrderForm'
+import {name as buyTicket} from '../buyTicket/buyTicket'
 import {name as mapLocation} from '../mapLocation/mapLocation'
+import {name as eventAttendees} from '../eventAttendees/eventAttendees'
 
 import templateUrl from './eventDetails.html'
 import './style.scss'
@@ -10,18 +11,33 @@ import './style.scss'
 import { Events } from '../../../api/events';
 
 class EventDetailsController {
-  constructor($scope, $reactive, $stateParams) {
+  constructor($scope, $reactive, $stateParams, $mdMedia) {
     'ngInject';
 
     $reactive(this).attach($scope);
 
     this.subscribe('event', () => [$stateParams.eventId])
+    
+    if ($stateParams.buy === "1") {
+      this.eventView = 'buyTicket';
+    }
 
     this.helpers({
       event() {
         return Events.findOne($stateParams.eventId)
       }
     })
+    
+    $scope.$watch(function() {
+      return $mdMedia('xs')
+    }, (xs) => {
+      this.xs = xs; // true
+    })
+  }
+  
+  onSuccess(eventView) {
+    console.log('eventDetails onSuccess called')
+    this.eventView = eventView
   }
 
 }
@@ -32,8 +48,9 @@ const name = 'eventDetails';
 export default angular.module(name, [
   'angular-meteor',
   'remarkable',
-  eventOrderForm,
+  buyTicket,
   mapLocation,
+  eventAttendees,
   uiRouter,
   ngMaterial
 ]).component(name, {
@@ -44,11 +61,11 @@ export default angular.module(name, [
 
 function config($stateProvider, $mdThemingProvider) {
   'ngInject';
-  $mdThemingProvider.theme('darkTheme').dark()
+  $mdThemingProvider.theme('darkTheme').primaryPalette('light-green').dark()
 
   $stateProvider
     .state('event', {
-      url: '/event/:eventId',
+      url: '/event/:eventId?:buy',
       template: '<event-details></event-details>'
     });
 }