+++ server/kadira.js
@@ -1,1 +1,4 @@
-// Kadira.connect('wuQgJELLDk92A2r99', '7274382d-5ebb-4492-95d3-c408090b8473');
+// // Kadira.connect('wuQgJELLDk92A2r99', '7274382d-5ebb-4492-95d3-c408090b8473');
+// if (process.env.METEOR_ENVIRONMENT == 'production') {
+//   Kadira.connect('BLAZSFbWJe2vZdnWk', '0febb9e8-4887-454f-9cb6-37042a76aa30');
+// }