+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.1.9 (27-6-2016)
+- added registration welcome email template
+- fixed admin products list being difficult to use
+- added ssr pages for facebook bots to get nice rendered products, events and store (users coming soon)
+- fix quick-link to profile from edit me page
+- server connection status component added and tested
+- basic events components completed
+  - admin can create events
+  - members of the public can read about event and buy tickets
+  - members get an email with tickets
+  - event can be shared on facebook and looks nice
+  - email drafted for reminder of the event the day it happens
+
+
+
 ## 0.1.8 (2-6-2016)
 - added colourful statuses to customer's sales and purchases like the admin has. This feature may need some refining in the near future.
 - added facebook link to navigation